renaming file for clarity, removing scroll-depth
